refactor: 장바구니 쿠폰 UI 분리
